thermal: int3400: Remove unneeded data_vault attribute_group
The group only contains a single entry and the conditionals around its
lifecycle make clear that this won't change.

Remove the unnecessary group.

This saves some memory and it's easier to read.

The removal of a non-const bin_attribute[] instance is also a
preparation for the constification of struct bin_attributes.

Signed-off-by: Thomas Weißschuh <linux@weissschuh.net>
Link: https://patch.msgid.link/20241121-sysfs-const-bin_attr-int340x_thermal-v1-1-2436facf9dae@weissschuh.net
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>